requirements.txt
beartype==0.18.5
colorama==0.4.6
colorspacious==1.1.2
dacite==1.8.1
e3nn==0.5.1
einops==0.8.0
hydra-core==1.3.2
jaxtyping==0.2.33
lpips==0.1.4
matplotlib==3.9.1
moviepy==1.0.3
numpy==1.24.4
opencv_python==4.11.0.86
Pillow==10.4.0
plyfile==1.1
pytorch_lightning==2.4.0
scikit-image==0.24.0
sk-video==1.1.10
tabulate==0.9.0
tqdm==4.66.4
wandb==0.17.7
git+https://github.com/dcharatan/diff-gaussian-rasterization-modified